msm: clock-debug: Enable debug muxes en route to clock being measured
MSM clock debug muxes have gates that need to be enabled
to allow clock signals to be output to the measurement
circuitry. Since we want to ensure that a clock that is
assumed to be off in software is also not outputting a
signal, it is not possible to call clk_enable on the
measure clock. Therefore call the enable op on all the
clocks upstream of the measure clock, except for the clock
being measured. Also, leave the clocks enabled to allow
measurement using a scope.

Change-Id: Ia8016edd460d03b6a0686c58788457aeb6c9e2c5
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>