i2c: qup: selective bus recovery
In multimaster configurations we should not attempt to recover
bus just because we got a timeout. First, timeout may happen
for reasons that have nothing to do with I2C, and there may be
another master talking.

Check the bus status and only do recovery when we're the master,
no one else claims the mastership and there is some i2c error
actually indicated in the status.

Also remove pointless check for driver status in the IRQ handler
since i2c interrupts are only enabled inside the xfer and we
always have clocks on at that time.

Signed-off-by: Igor Kovalenko <cik009@motorola.com>
Reviewed-by: Igor Kovalenko <igork@motorola.com>