i2c-qup: Add 1-clk pulse recovery
Port of (CR).

If the transaction is corrupted in such a way that the host
and device are out-of-sync, then the device can output
it's ACK signal indefinitely, and the 9-cycle recovery method
will clear the ACK on the first cycle, and then force a new
ACK on the 9th cycle, allowing the failure to continue with
DATA driven low.To fix this issue the 1-cycle recovery method
added to recover the bus, after the 9-cycle recovery method failed.
This is a workaround for specific LM3642 HW problem and may not
work with other slave devices. To enable the workaround set
qcom,extended-recovery device tree flag to 1.

Signed-off-by: Vladimir Karpovich <cvk011c@motorola.com>
Reviewed-by: Igor Kovalenko <igork@motorola.com>
Reviewed-by: Lian-Wei Wang <lian-wei.wang@motorola.com>