HOTPLUGS: enabled CPU 0-2 on screen OFF so we have 3 cores when in standby.
Default LP is all cores online in standby. when suspending and no wakelocks.
only boot CPU is online cpu0 and sleeping.

i have found that when we have only core0 and on limited freq.
and play hifi sound via stream wifi/data cpu is 100% on core0
and other system services start to die. and kernel panic on mem allocations
and OOM and watchdog barks...

Give system what it's needs.
now load is split per core, and system is relaxed.